+++ Code/Editor/TrackView/TrackViewDopeSheetBase.cpp
@@ -258,7 +258,7 @@ void CTrackViewDopeSheetBase::SetTimeScale(float timeScale, float fAnchorTime)
     while (fPixelsPerTick >= 12.0 && steps < 100);
 
     float fCurrentOffset = -fAnchorTime * m_timeScale;
-    m_scrollOffset.rx() += fOldOffset - fCurrentOffset;
+    m_scrollOffset.rx() += static_cast<int>(fOldOffset - fCurrentOffset);
     m_scrollBar->setValue(m_scrollOffset.x());
 
     update();
